moved unconstrained distribution to a dedicated module
